Flowers: Add black tulip, green chrysanthemum
Adds a direct resource of green dye, and a flower resource of black dye.
Completes the colour sequence: red, orange, yellow, green, blue, violet.
Makes all base dyes cultivatable and sustainable, without the presence
of coal.

Add the new flowers to the world with the same density variation as the
others, but obviously with different noise seeds. Results in more flowers
in a world and more variety of flower combinations.